requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-telegram-bot==22.5
soupsieve==2.8.1
typing_extensions==4.15.0
Werkzeug==3.1.5
